// src/lib/db/schema.ts
export const emailVerificationTokens = mysqlTable(
  'email_verification_tokens',
  {
    id: serial('id').primaryKey().autoincrement(),
    userId: bigint('user_id', {
      mode: 'number',
    }).notNull(),
    email: varchar('email', { length: 319 }).notNull(),
    token: varchar('token', { length: 255 }).notNull().unique(),
    expiresAt: timestamp('expires_at').notNull(),
    createdAt: timestamp('created_at').defaultNow(),
  },
  (emailVerificationTokens) => ({
    tokenIndex: uniqueIndex('token_idx').on(emailVerificationTokens.token),
    emailIndex: uniqueIndex('email_idx').on(emailVerificationTokens.email),
  })
);

export const sessions = mysqlTable('sessions', {
  id: serial('id').primaryKey(),
  userId: bigint('user_id', {
    mode: 'number',
  }).notNull(),
  sessionToken: varchar('session_token', { length: 255 }).notNull(),
  expiresAt: timestamp('expires_at').notNull(),
  createdAt: timestamp('created_at').defaultNow(),
});

export const users = mysqlTable(
  'users',
  {
    id: serial('id').primaryKey().autoincrement(),
    publicId: varchar('public_id', { length: 12 }).notNull(),
    email: varchar('email', { length: 319 }).notNull().unique(),
    createdAt: timestamp('created_at').defaultNow(),
    updatedAt: timestamp('updated_at').defaultNow(),
    /**
     * The most efficient MySQL type for a bcrypt password hash is varchar(60).
     */
    passwordHash: varchar('password_hash', {
      length: 200,
    }),
    emailVerifiedAt: timestamp('email_verified_at'),
  },
  (users) => ({
    emailIndex: uniqueIndex('email_idx').on(users.email),
    publicId: uniqueIndex('public_id_idx').on(users.publicId),
  })
);